+++ saleor/plugins/avatax/plugin.py
@@ -141,12 +141,8 @@ class AvataxPlugin(BasePlugin):
             if line_info.variant.product.charge_taxes:
                 continue
             line_price = base_calculations.base_checkout_line_total(
-                line_info.line,
-                line_info.variant,
-                line_info.product,
-                line_info.collections,
+                line_info,
                 channel,
-                line_info.channel_listing,
                 discounts,
             )
             price.gross.amount += line_price.gross.amount
@@ -163,7 +159,6 @@ class AvataxPlugin(BasePlugin):
     ) -> TaxedMoney:
         if self._skip_plugin(previous_value):
             return previous_value
-
         checkout_total = previous_value
 
         if not _validate_checkout(checkout, [line_info.line for line_info in lines]):
@@ -338,24 +333,20 @@ class AvataxPlugin(BasePlugin):
     def calculate_checkout_line_total(
         self,
         checkout: "Checkout",
-        checkout_line: "CheckoutLine",
-        variant: "ProductVariant",
-        product: "Product",
-        collections: Iterable["Collection"],
+        checkout_line_info: "CheckoutLineInfo",
         address: Optional["Address"],
         channel: "Channel",
-        channel_listing: "ProductVariantChannelListing",
-        discounts: Iterable[DiscountInfo],
+        discounts: Iterable["DiscountInfo"],
         previous_value: TaxedMoney,
     ) -> TaxedMoney:
         if self._skip_plugin(previous_value):
             return previous_value
 
         base_total = previous_value
-        if not checkout_line.variant.product.charge_taxes:
+        if not checkout_line_info.product.charge_taxes:
             return base_total
 
-        if not _validate_checkout(checkout, [checkout_line]):
+        if not _validate_checkout(checkout, [checkout_line_info.line]):
             return base_total
 
         taxes_data = get_checkout_tax_data(checkout, discounts, self.config)
@@ -364,7 +355,7 @@ class AvataxPlugin(BasePlugin):
 
         currency = taxes_data.get("currencyCode")
         for line in taxes_data.get("lines", []):
-            if line.get("itemCode") == variant.sku:
+            if line.get("itemCode") == checkout_line_info.variant.sku:
                 tax = Decimal(line.get("tax", 0.0))
                 line_net = Decimal(line["lineAmount"])
                 line_gross = Money(amount=line_net + tax, currency=currency)
@@ -373,29 +364,67 @@ class AvataxPlugin(BasePlugin):
 
         return base_total
 
-    def _calculate_order_line_unit(self, order_line):
-        order = order_line.order
-        taxes_data = get_order_tax_data(order, self.config)
+    def calculate_checkout_line_unit_price(
+        self,
+        checkout: "Checkout",
+        checkout_line_info: "CheckoutLineInfo",
+        address: Optional["Address"],
+        discounts: Iterable["DiscountInfo"],
+        channel: "Channel",
+        previous_value: TaxedMoney,
+    ):
+        if not checkout_line_info.product.charge_taxes:
+            return previous_value
+        return self._calculate_unit_price(
+            checkout,
+            checkout_line_info.line,
+            checkout_line_info.variant,
+            previous_value,
+            discounts,
+            is_order=False,
+        )
+
+    def calculate_order_line_unit(
+        self,
+        order: "Order",
+        order_line: "OrderLine",
+        variant: "ProductVariant",
+        product: "Product",
+        previous_value: TaxedMoney,
+    ) -> TaxedMoney:
+        if not variant or (variant and not product.charge_taxes):
+            return previous_value
+        return self._calculate_unit_price(
+            order, order_line, variant, previous_value, is_order=True
+        )
+
+    def _calculate_unit_price(
+        self,
+        instance: Union["Checkout", "Order"],
+        line: Union["CheckoutLine", "OrderLine"],
+        variant: "ProductVariant",
+        base_value: TaxedMoney,
+        discounts: Optional[Iterable[DiscountInfo]] = [],
+        *,
+        is_order: bool,
+    ):
+        lines = [] if is_order else [line]
+        taxes_data = self._get_tax_data(
+            instance, base_value, is_order, discounts, lines  # type: ignore
+        )
+        if taxes_data is None:
+            return base_value
         currency = taxes_data.get("currencyCode")
-        for line in taxes_data.get("lines", []):
-            if line.get("itemCode") == order_line.variant.sku:
-                tax = Decimal(line.get("tax", 0.0)) / order_line.quantity
-                net = Decimal(line.get("lineAmount", 0.0)) / order_line.quantity
+        for line_data in taxes_data.get("lines", []):
+            if line_data.get("itemCode") == variant.sku:
+                tax = Decimal(line_data.get("tax", 0.0)) / line.quantity
+                net = Decimal(line_data.get("lineAmount", 0.0)) / line.quantity
 
                 gross = Money(amount=net + tax, currency=currency)
                 net = Money(amount=net, currency=currency)
                 return TaxedMoney(net=net, gross=gross)
 
-    def calculate_order_line_unit(
-        self, order_line: "OrderLine", previous_value: TaxedMoney
-    ) -> TaxedMoney:
-        if self._skip_plugin(previous_value):
-            return previous_value
-        if order_line.variant and not order_line.variant.product.charge_taxes:  # type: ignore
-            return previous_value
-        if _validate_order(order_line.order):
-            return self._calculate_order_line_unit(order_line)
-        return order_line.unit_price
+        return base_value
 
     def calculate_order_shipping(
         self, order: "Order", previous_value: TaxedMoney
@@ -515,12 +544,12 @@ class AvataxPlugin(BasePlugin):
     def _get_tax_data(
         self,
         instance: Union["Order", "Checkout"],
-        base_rate: Decimal,
+        base_value: Decimal,
         is_order: bool,
         discounts: Optional[Iterable[DiscountInfo]] = None,
         checkout_lines: Iterable["CheckoutLine"] = [],
     ):
-        if self._skip_plugin(base_rate):
+        if self._skip_plugin(base_value):
             return None
 
         valid = (